+++ src/pyguitest/__init__.py
@@ -66,7 +66,7 @@ if TYPE_CHECKING:
 
 _T = TypeVar("_T")
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 __all__ = [
     "connect",